+++ NEWS.md
@@ -37,6 +37,8 @@ Compiler/Runtime improvements
 Deprecated or removed
 ---------------------
 
+  * `isdefined(a::Array, i::Int)` has been deprecated in favor of `isassigned` ([#18346]).
+
 Julia v0.5.0 Release Notes
 ==========================
 
@@ -653,3 +655,4 @@ Language tooling improvements
 [#17668]: https://github.com/JuliaLang/julia/issues/17668
 [#17785]: https://github.com/JuliaLang/julia/issues/17785
 [#18330]: https://github.com/JuliaLang/julia/issues/18330
+[#18346]: https://github.com/JuliaLang/julia/issues/18346